requirements.txt

cryptography
fastapi
mss
pillow
pynput
pyautogui
uvicorn
websockets
opencv-python
PyJWT
bcrypt
psutil
browserhistory
pytest
pytest-asyncio
anyio
# for local development
python-dotenv